Move NewFromText to yaml.go
